added setup snippet in App.css, has sr-only class, also created function for svg files in other component called Arrows.js
